+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 
 - A `Count rows` button is displayed in table row-list buffers.
 
+- The connection widget uses initial values for database name, username, hostname and password taken
+  from environment variables such as `POSTGRES_DATABASE` and `POSTGRES_USER`, if defined.
+
 
 ## [0.7] - 2024-05-23